requirements.txt
asgiref==3.5.1
attrs==21.4.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==37.0.2
distro-info==1.0
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
django-rest-jwt==0.1.46
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
drf-nested-routers==0.93.4
iniconfig==1.1.1
install==1.3.5
jwt==1.3.1
packaging==21.3
pluggy==1.0.0
psycopg==3.0.12
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.8
pytest==7.0.0
pytest-django==4.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
